requirements.txt
playwright
beautifulsoup4
requests
psycopg2-binary
Pillow
tqdm
loguru